rename deploy/command.ts to deploy/spec.ts
