+++ packages/cli/src/commands/deploy/spec.ts
